pyproject.toml
[project]
name = "choresplanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.129.0",
    "google-api-python-client>=2.189.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "inflect>=7.5.0",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "rich>=14.3.2",
    "sqlalchemy>=2.0.46",
    "tortoise-orm[ipython]>=1.1.5",
    "typer>=0.21.1",
]

[dependency-groups]
dev = [
    "ipython>=9.10.0",
]
test = [
    "freezegun>=1.5.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]

[tool.tortoise]
tortoise_orm = "src.db.TORTOISE_ORM"

